Refactor CsvRowArtifact to inherit from BaseTextArtifact
