+++ tests/unit/artifacts/test_list_artifact.py
@@ -1,6 +1,6 @@
 import pytest
 
-from griptape.artifacts import BlobArtifact, CsvRowArtifact, ListArtifact, TextArtifact
+from griptape.artifacts import BaseTextArtifact, BlobArtifact, CsvRowArtifact, ListArtifact, TextArtifact
 
 
 class TestListArtifact:
@@ -32,7 +32,7 @@ class TestListArtifact:
 
     def test_is_type(self):
         assert ListArtifact([TextArtifact("foo")]).is_type(TextArtifact)
-        assert ListArtifact([CsvRowArtifact({"foo": "bar"})]).is_type(TextArtifact)
+        assert ListArtifact([CsvRowArtifact({"foo": "bar"})]).is_type(BaseTextArtifact)
         assert ListArtifact([CsvRowArtifact({"foo": "bar"})]).is_type(CsvRowArtifact)
 
     def test_has_items(self):
